JIRA-5: Introduce the full concept of Users API
This commit is very huge. As I wrote in the README.md, I had to
use it for start. I will build up the products/reviews API on top
of this concept.
I tried to follow everything slowly, in order to understand it.
Regarding tests, it's still in TODO state, I had struggled a lot
with configuring them. In this commit, there is only an example
how it could be done.